// src/app/dashboard-page/Models/clientcust.model.ts
export class Clientcust {
  id: number;
  name: string;
  lastname: string;
  birthdate: string;
  cpf: string;
  rg: string;
  phone: string;
  landline: string;
  address: string;
  house_number: string;
  district: string;
  city: string;
  state: string;
  created_at: string;
  updated_at: string;
}